+++ src/main/java/com/ormi/mogakcote/notice/infrastructure/NoticeRepository.java
@@ -2,6 +2,7 @@ package com.ormi.mogakcote.notice.infrastructure;
 
 import com.ormi.mogakcote.notice.domain.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -13,4 +14,7 @@ import java.util.Optional;
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
     Optional<Notice> findByNoticeId(Long noticeId);
     List<Notice> findTop5ByOrderByCreatedAtDesc();
+
+    @Query("SELECT n FROM Notice n ORDER BY n.createdAt DESC LIMIT 5")
+    List<Notice> getNoticeLatestFive();
 }
