+++ src/main/java/com/ormi/mogakcote/post/dto/response/PostSearchResponse.java
@@ -1,0 +1,20 @@
+package com.ormi.mogakcote.post.dto.response;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class PostSearchResponse {
+	private Long id;
+	private String title;
+	private String content;
+	private List<String> algorithms;
+	private Integer viewCnt;
+	private LocalDateTime createdAt;
+}
