+++ src/main/java/com/ormi/mogakcote/post/presentation/PostController.java
@@ -1,6 +1,4 @@
-package com.ormi.mogakcote.post;
-
-import static com.ormi.mogakcote.post.SortType.*;
+package com.ormi.mogakcote.post.presentation;
 
 import java.util.HashMap;
 import java.util.List;
@@ -10,14 +8,16 @@ import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ormi.mogakcote.auth.model.AuthUser;
 import com.ormi.mogakcote.common.model.ResponseDto;
+import com.ormi.mogakcote.notice.dto.response.NoticeResponse;
+import com.ormi.mogakcote.post.application.PostService;
+import com.ormi.mogakcote.post.dto.request.PostSearchRequest;
+import com.ormi.mogakcote.post.dto.response.PostSearchResponse;
 
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -32,7 +32,7 @@ public class PostController {
 		 @ModelAttribute PostSearchRequest postSearchRequest
 	) {
 		List<NoticeResponse> noticeResponse = postService.getNoticeLatestFive();
-		Page<PostResponse> postResponse = postService.searchPost(user, postSearchRequest);
+		Page<PostSearchResponse> postResponse = postService.searchPost(user, postSearchRequest);
 
 		Map<String, Object> map = new HashMap<>();
 		map.put("notice", noticeResponse);
